move validate_source_ds to utils.io and the coresponding test to test_utils_io.py
